nakshatralu tab,audio,video added delete files from folder updated

+++ delete-grahalu.php
@@ -11,9 +11,16 @@ $db->connect();
 		exit(0);
 	}
 	$data = array();
-
-	$sql_query = "DELETE  FROM grahalu WHERE id =" . $ID;
+	$sql_query = "SELECT *  FROM grahalu WHERE id =" . $ID;
 	$db->sql($sql_query);
 	$res = $db->getResult();
-	header("location:grahalu.php");
+	$target_path = "".$res[0]['image'];
+		if(unlink($target_path)){	
+
+				$sql_query = "DELETE  FROM grahalu WHERE id =" . $ID;
+				$db->sql($sql_query);
+				$res = $db->getResult();
+				header("location:grahalu.php");
+
+		}
 ?>
