+++ delete-poojalu_submenu.php
@@ -11,9 +11,14 @@ $db->connect();
 		exit(0);
 	}
 	$data = array();
-
-	$sql_query = "DELETE  FROM poojalu_submenu WHERE id =" . $ID;
+	$sql_query = "SELECT *  FROM poojalu_submenu WHERE id =" . $ID;
 	$db->sql($sql_query);
 	$res = $db->getResult();
-	header("location:poojalu_submenu.php");
+	$target_path = "".$res[0]['image'];
+		if(unlink($target_path)){	
+					$sql_query = "DELETE  FROM poojalu_submenu WHERE id =" . $ID;
+					$db->sql($sql_query);
+					$res = $db->getResult();
+					header("location:poojalu_submenu.php");
+		}
 ?>
